+++ src/PhpWord/Element/TextRun.php
@@ -28,8 +28,6 @@ class TextRun extends Container
      * Create new instance
      *
      * @param string|array|Paragraph $paragraphStyle
-     * @param string $docPart section|header|footer
-     * @param int $docPartId
      */
     public function __construct($paragraphStyle = null)
     {