+++ src/PhpWord/Writer/Word2007.php
@@ -29,18 +29,18 @@ use PhpOffice\PhpWord\Writer\Word2007\Styles;
 class Word2007 extends Writer implements IWriter
 {
     /**
-     * Types of images
+     * Content types values
      *
      * @var array
      */
-    private $imageTypes = array();
+    private $cTypes = array('default' => array(), 'override' => array());
 
     /**
-     * Types of objects
+     * Document relationship
      *
      * @var array
      */
-    private $objectTypes = array();
+    private $docRels = array();
 
     /**
      * Create new Word2007 writer
@@ -98,14 +98,18 @@ class Word2007 extends Writer implements IWriter
                 }
             }
 
-            // Add section media files
-            $sectionElements = array();
+            // Content types
+            $this->cTypes['default'] = array(
+                'rels' => 'application/vnd.openxmlformats-package.relationships+xml',
+                'xml'  => 'application/xml',
+            );
 
-            $secElements = Media::getMediaElements('section');
-            if (!empty($secElements)) {
-                $this->addFilesToPackage($objZip, $secElements);
-                foreach ($secElements as $element) {
-                    $sectionElements[] = $element;
+            // Add section media files
+            $sectionMedia = Media::getElements('section');
+            if (!empty($sectionMedia)) {
+                $this->addFilesToPackage($objZip, $sectionMedia);
+                foreach ($sectionMedia as $element) {
+                    $this->docRels[] = $element;
                 }
             }
 
@@ -114,70 +118,32 @@ class Word2007 extends Writer implements IWriter
             $this->addHeaderFooterMedia($objZip, 'footer');
 
             // Add header/footer contents
-            $cHdrs = 0;
-            $cFtrs = 0;
-            $rID = Media::countMediaElements('section') + 6; // @see Rels::writeDocRels for 6 first elements
+            $overrides = array();
+            $rID = Media::countElements('section') + 6; // @see Rels::writeDocRels for 6 first elements
             $sections = $this->phpWord->getSections();
-            $footers = array();
             foreach ($sections as $section) {
-                $headers = $section->getHeaders();
-                if (!empty($headers)) {
-                    foreach ($headers as $index => &$header) {
-                        $cHdrs++;
-                        $header->setRelationId(++$rID);
-                        $hdrFile = "header{$cHdrs}.xml";
-                        $sectionElements[] = array('target' => $hdrFile, 'type' => 'header', 'rID' => $rID);
-                        $objZip->addFromString(
-                            "word/{$hdrFile}",
-                            $this->getWriterPart('header')->writeHeader($header)
-                        );
-                    }
-                }
-                $footer = $section->getFooter();
-                $footers[++$cFtrs] = $footer;
-                if (!is_null($footer)) {
-                    $footer->setRelationId(++$rID);
-                    $footerCount = $footer->getSectionId();
-                    $ftrFile = "footer{$footerCount}.xml";
-                    $sectionElements[] = array('target' => $ftrFile, 'type' => 'footer', 'rID' => $rID);
-                    $objZip->addFromString(
-                        "word/{$ftrFile}",
-                        $this->getWriterPart('footer')->writeFooter($footer)
-                    );
-                }
+                $this->addHeaderFooterContent($section, $objZip, 'header', $rID);
+                $this->addHeaderFooterContent($section, $objZip, 'footer', $rID);
             }
 
             // Add footnotes media files, relations, and contents
             if (Footnote::countFootnoteElements() > 0) {
-                $sectionElements[] = array('target' => 'footnotes.xml', 'type' => 'footnotes', 'rID' => ++$rID);
-                $footnoteMedia = Media::getMediaElements('footnote');
+                $footnoteMedia = Media::getElements('footnote');
                 $this->addFilesToPackage($objZip, $footnoteMedia);
                 if (!empty($footnoteMedia)) {
-                    $objZip->addFromString(
-                        'word/_rels/footnotes.xml.rels',
-                        $this->getWriterPart('rels')->writeMediaRels($footnoteMedia)
-                    );
+                    $objZip->addFromString('word/_rels/footnotes.xml.rels', $this->getWriterPart('rels')->writeMediaRels($footnoteMedia));
                 }
-                $objZip->addFromString(
-                    'word/footnotes.xml',
-                    $this->getWriterPart('footnotes')->writeFootnotes(Footnote::getFootnoteElements())
-                );
+                $objZip->addFromString('word/footnotes.xml', $this->getWriterPart('footnotes')->writeFootnotes(Footnote::getFootnoteElements()));
+                $this->cTypes['override']["/word/footnotes.xml"] = 'footnotes';
+                $this->docRels[] = array('target' => 'footnotes.xml', 'type' => 'footnotes', 'rID' => ++$rID);
             }
 
             // Write dynamic files
-            $objZip->addFromString(
-                '[Content_Types].xml',
-                $this->getWriterPart('contenttypes')->writeContentTypes(
-                    $this->imageTypes,
-                    $this->objectTypes,
-                    $cHdrs,
-                    $footers
-                )
-            );
+            $objZip->addFromString('[Content_Types].xml', $this->getWriterPart('contenttypes')->writeContentTypes($this->cTypes));
             $objZip->addFromString('_rels/.rels', $this->getWriterPart('rels')->writeMainRels());
             $objZip->addFromString('docProps/app.xml', $this->getWriterPart('docprops')->writeDocPropsApp($this->phpWord));
             $objZip->addFromString('docProps/core.xml', $this->getWriterPart('docprops')->writeDocPropsCore($this->phpWord));
-            $objZip->addFromString('word/_rels/document.xml.rels', $this->getWriterPart('rels')->writeDocRels($sectionElements));
+            $objZip->addFromString('word/_rels/document.xml.rels', $this->getWriterPart('rels')->writeDocRels($this->docRels));
             $objZip->addFromString('word/document.xml', $this->getWriterPart('document')->writeDocument($this->phpWord));
             $objZip->addFromString('word/styles.xml', $this->getWriterPart('styles')->writeStyles($this->phpWord));
 
@@ -200,54 +166,7 @@ class Word2007 extends Writer implements IWriter
     }
 
     /**
-     * Check content types
-     *
-     * @param string $src
-     */
-    private function checkContentTypes($src)
-    {
-        $extension = null;
-        if (stripos(strrev($src), strrev('.php')) === 0) {
-            $extension = 'php';
-        } else {
-            if (function_exists('exif_imagetype')) {
-                $imageType = exif_imagetype($src);
-            } else {
-                $tmp = getimagesize($src);
-                $imageType = $tmp[2];
-            }
-            if ($imageType === \IMAGETYPE_JPEG) {
-                $extension = 'jpg';
-            } elseif ($imageType === \IMAGETYPE_GIF) {
-                $extension = 'gif';
-            } elseif ($imageType === \IMAGETYPE_PNG) {
-                $extension = 'png';
-            } elseif ($imageType === \IMAGETYPE_BMP) {
-                $extension = 'bmp';
-            } elseif ($imageType === \IMAGETYPE_TIFF_II || $imageType === \IMAGETYPE_TIFF_MM) {
-                $extension = 'tif';
-            }
-        }
-
-        if (isset($extension)) {
-            $imageData = getimagesize($src);
-            $imageType = image_type_to_mime_type($imageData[2]);
-            $imageExtension = str_replace('.', '', image_type_to_extension($imageData[2]));
-            if ($imageExtension === 'jpeg') {
-                $imageExtension = 'jpg';
-            }
-            if (!in_array($imageType, $this->imageTypes)) {
-                $this->imageTypes[$imageExtension] = $imageType;
-            }
-        } else {
-            if (!in_array($extension, $this->objectTypes)) {
-                $this->objectTypes[] = $extension;
-            }
-        }
-    }
-
-    /**
-     * Check content types
+     * Add section files to package
      *
      * @param mixed $objZip
      * @param mixed $elements
@@ -255,47 +174,78 @@ class Word2007 extends Writer implements IWriter
     private function addFilesToPackage($objZip, $elements)
     {
         foreach ($elements as $element) {
+            // Do not add link
             if ($element['type'] == 'link') {
                 continue;
             }
+            // Retrieve remote image
             if (isset($element['isMemImage']) && $element['isMemImage']) {
-                $image = call_user_func($element['createfunction'], $element['source']);
+                $image = call_user_func($element['createFunction'], $element['source']);
                 ob_start();
-                call_user_func($element['imagefunction'], $image);
+                call_user_func($element['imageFunction'], $image);
                 $imageContents = ob_get_contents();
                 ob_end_clean();
                 $objZip->addFromString('word/' . $element['target'], $imageContents);
                 imagedestroy($image);
-
-                $this->checkContentTypes($element['source']);
             } else {
                 $objZip->addFile($element['source'], 'word/' . $element['target']);
-                $this->checkContentTypes($element['source']);
+            }
+            // Register content types
+            if ($element['type'] == 'image') {
+                $imageExtension = $element['imageExtension'];
+                $imageType = $element['imageType'];
+                if (!array_key_exists($imageExtension, $this->cTypes['default'])) {
+                    $this->cTypes['default'][$imageExtension] = $imageType;
+                }
+            } else {
+                if (!array_key_exists('bin', $this->cTypes['default'])) {
+                    $this->cTypes['default']['bin'] = 'application/vnd.openxmlformats-officedocument.oleObject';
+                }
             }
         }
     }
 
     /**
-     * Add header/footer media elements
+     * Add header/footer media files
      *
      * @param mixed $objZip
      * @param string $docPart
      */
     private function addHeaderFooterMedia($objZip, $docPart)
     {
-        $elements = Media::getMediaElements($docPart);
+        $elements = Media::getElements($docPart);
         if (!empty($elements)) {
             foreach ($elements as $file => $media) {
                 if (count($media) > 0) {
-                    $objZip->addFromString(
-                        'word/_rels/' . $file . '.xml.rels',
-                        $this->getWriterPart('rels')->writeMediaRels($media)
-                    );
                     if (!empty($media)) {
                         $this->addFilesToPackage($objZip, $media);
                     }
+                    $objZip->addFromString("word/_rels/{$file}.xml.rels", $this->getWriterPart('rels')->writeMediaRels($media));
                 }
             }
+        }
+    }
+
+    /**
+     * Add header/footer content
+     *
+     * @param Section $section
+     * @param string $elmType
+     * @param integer $rID
+     */
+    private function addHeaderFooterContent(&$section, $objZip, $elmType, &$rID)
+    {
+        $getFunction = $elmType == 'header' ? 'getHeaders' : 'getFooters';
+        $writeFunction = $elmType == 'header' ? 'writeHeader' : 'writeFooter';
+        $elmCount = ($section->getSectionId() - 1) * 3;
+        $elmObjects = $section->$getFunction();
+        foreach ($elmObjects as $index => &$elmObject) {
+            $elmCount++;
+            $elmObject->setRelationId(++$rID);
+            $elmFile = "{$elmType}{$elmCount}.xml";
+            $objZip->addFromString("word/$elmFile", $this->getWriterPart($elmType)->$writeFunction($elmObject));
+            $this->cTypes['override']["/word/$elmFile"] = $elmType;
+            $this->docRels[] = array('target' => $elmFile, 'type' => $elmType, 'rID' => $rID);
         }
     }
 }