+++ src/PhpWord/Writer/Word2007/ContentTypes.php
@@ -19,28 +19,13 @@ class ContentTypes extends WriterPart
 {
     /**
      * Write [Content_Types].xml
-     * @param array $imageTypes
-     * @param array $objectTypes
-     * @param int $cHdrs
-     * @param array $footers
+     *
+     * @param array $contentTypes
      */
-    public function writeContentTypes($imageTypes, $objectTypes, $cHdrs, $footers)
+    public function writeContentTypes($contentTypes)
     {
-
         $OpenXMLPrefix = 'application/vnd.openxmlformats-';
-        $WordMLPrefix = $OpenXMLPrefix . 'officedocument.wordprocessingml.';
-
-        $defaults = array(
-            'rels' => $OpenXMLPrefix . 'package.relationships+xml',
-            'xml'  => 'application/xml',
-
-        );
-        if (is_array($imageTypes)) {
-            $defaults = array_merge($defaults, $imageTypes);
-        }
-        if (count($objectTypes) > 0) {
-            $defaults['bin'] = $OpenXMLPrefix . 'officedocument.oleObject';
-        }
+        $WordMLPrefix  = $OpenXMLPrefix . 'officedocument.wordprocessingml.';
         $overrides = array(
             '/docProps/core.xml'     => $OpenXMLPrefix . 'package.core-properties+xml',
             '/docProps/app.xml'      => $OpenXMLPrefix . 'officedocument.extended-properties+xml',
@@ -51,14 +36,12 @@ class ContentTypes extends WriterPart
             '/word/theme/theme1.xml' => $OpenXMLPrefix . 'officedocument.theme+xml',
             '/word/webSettings.xml'  => $WordMLPrefix  . 'webSettings+xml',
             '/word/fontTable.xml'    => $WordMLPrefix  . 'fontTable+xml',
-            '/word/footnotes.xml'    => $WordMLPrefix  . 'footnotes+xml',
         );
-        for ($i = 1; $i <= $cHdrs; $i++) {
-            $overrides["/word/header{$i}.xml"] = $WordMLPrefix  . 'header+xml';
-        }
-        for ($i = 1; $i <= count($footers); $i++) {
-            if (!is_null($footers[$i])) {
-                $overrides["/word/footer{$i}.xml"] = $WordMLPrefix  . 'footer+xml';
+
+        $defaults = $contentTypes['default'];
+        if (!empty($contentTypes['override'])) {
+            foreach ($contentTypes['override'] as $key => $val) {
+                $overrides[$key] = $WordMLPrefix . $val . '+xml';
             }
         }
 
@@ -66,12 +49,8 @@ class ContentTypes extends WriterPart
         $xmlWriter->startDocument('1.0', 'UTF-8', 'yes');
         $xmlWriter->startElement('Types');
         $xmlWriter->writeAttribute('xmlns', 'http://schemas.openxmlformats.org/package/2006/content-types');
-        foreach ($defaults as $key => $value) {
-            $this->writeContentType($xmlWriter, true, $key, $value);
-        }
-        foreach ($overrides as $key => $value) {
-            $this->writeContentType($xmlWriter, false, $key, $value);
-        }
+        $this->writeContentType($xmlWriter, $defaults, true);
+        $this->writeContentType($xmlWriter, $overrides, false);
         $xmlWriter->endElement();
 
         return $xmlWriter->getData();
@@ -86,62 +65,19 @@ class ContentTypes extends WriterPart
      * @param string $contentType Content type
      * @throws Exception
      */
-    private function writeContentType(XMLWriter $xmlWriter, $isDefault, $partName = '', $contentType = '')
+    private function writeContentType(XMLWriter $xmlWriter, $parts, $isDefault)
     {
-        if ($partName != '' && $contentType != '') {
-            $element = $isDefault ? 'Default' : 'Override';
-            $partAttribute = $isDefault ? 'Extension' : 'PartName';
-            $xmlWriter->startElement($element);
-            $xmlWriter->writeAttribute($partAttribute, $partName);
-            $xmlWriter->writeAttribute('ContentType', $contentType);
-            $xmlWriter->endElement();
-        } else {
-            throw new Exception("Invalid parameters passed.");
+        foreach ($parts as $partName => $contentType) {
+            if ($partName != '' && $contentType != '') {
+                $partType = $isDefault ? 'Default' : 'Override';
+                $partAttribute = $isDefault ? 'Extension' : 'PartName';
+                $xmlWriter->startElement($partType);
+                $xmlWriter->writeAttribute($partAttribute, $partName);
+                $xmlWriter->writeAttribute('ContentType', $contentType);
+                $xmlWriter->endElement();
+            } else {
+                throw new Exception("Invalid parameters passed.");
+            }
         }
-    }
-
-    /**
-     * Get image mime type
-     *
-     * @param  string $pFile Filename
-     * @return string Mime Type
-     * @throws Exception
-     */
-    private function getImageMimeType($pFile = '')
-    {
-        if (file_exists($pFile)) {
-            $image = getimagesize($pFile);
-            return image_type_to_mime_type($image[2]);
-        } else {
-            throw new Exception("File $pFile does not exist");
-        }
-    }
-
-    /**
-     * Write Default XML element
-     *
-     * @param  XMLWriter $xmlWriter
-     * @param  string $partName Part name
-     * @param  string $contentType Content type
-     * @deprecated 0.9.2
-     * @codeCoverageIgnore
-     */
-    private function writeDefaultContentType(XMLWriter $xmlWriter, $partName = '', $contentType = '')
-    {
-        $this->writeContentType($xmlWriter, true, $partName, $contentType);
-    }
-
-    /**
-     * Write Override XML element
-     *
-     * @param  XMLWriter $xmlWriter
-     * @param  string $partName Part name
-     * @param  string $contentType Content type
-     * @deprecated 0.9.2
-     * @codeCoverageIgnore
-     */
-    private function writeOverrideContentType(XMLWriter $xmlWriter, $partName = '', $contentType = '')
-    {
-        $this->writeContentType($xmlWriter, false, $partName, $contentType);
     }
 }